Extract Token into its own source file
